day3-bai1

+++ day 3/bai1.py
@@ -1,5 +1,5 @@
-number1 = int(input("number1:"))
-number2 = int(input("number2:"))
+number1 = float(input("number1:"))
+number2 = float(input("number2:"))
 print("tổng:", number1+number2)
 print("hiệu:", number1-number2)
 print("tích:", number1*number2)
